+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: nfascia <nathanfascia@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/24 11:33:05 by nfascia           #+#    #+#             */
-/*   Updated: 2021/11/24 11:33:23 by nfascia          ###   ########.fr       */
+/*   Updated: 2021/11/24 15:34:37 by nfascia          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	src_size;
+	unsigned long	i;
+	unsigned long	src_size;
 
 	i = 0;
 	src_size = ft_strlen((char *)src);
